+++ samples/tutorial/tutorial-base/src/main/java/workflow/tutorial/TodoModel.kt
@@ -1,0 +1,6 @@
+package workflow.tutorial
+
+data class TodoModel(
+  val title: String,
+  val note: String
+)
